3.0 GenAI/3.1 Langchain/LangChain for LLM Application Development/agent_app/requirements.txt
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.1.7
langchain-chroma>=0.2.5
chromadb
openai==1.30.1
python-dotenv==1.0.1
tiktoken==0.7.0
